+++ NEWS.md
@@ -4,6 +4,10 @@
 * `simulate_data()` simulates a complete and incomplete data set.
 * `count_NA()` counts the number of missing values in a vector, data frame or 
   matrix.
+  
+## Improvements
+* `apply_imputation()` and friends now have an option to convert tibbles 
+  instead of throwing an error.
 
 ## Fixes
 * `delete_MAR_1_to_x()` and `delete_MNAR_1_to_x()` now handle unordered factors